+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.12.0
 ## Boxes
 * imroved connection routing
+* UI got an own comment panel to improve the comment display
 
 # v0.11.0
 ## Boxes